requirements.txt
paho_mqtt==1.5.1
pyserial==3.4
requests==2.24.0
influxdb==5.3.0

